refactor(omptype): collapse dual validation engine into one walker
